Add Type and Thumbnail to Asset class

+++ BynderExtension/Workers/AssetDownloadWorker.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using Bynder.Api;
 using Bynder.Api.Model;
 using Bynder.Names;
@@ -37,6 +37,7 @@ namespace Bynder.Workers
 
             // download asset information
             Asset asset = _bynderClient.GetAssetByAssetId(bynderId);
+            
             if (asset == null)
             {
                 _inRiverContext.Log(LogLevel.Error, "Asset information is empty");
@@ -51,41 +52,55 @@ namespace Bynder.Workers
             var resourceFileId = resourceEntity.GetField(FieldTypeIds.ResourceFileId)?.Data;
             var existingFileId = (int?)resourceFileId ?? 0;
 
-            // download the asset
-            var file = _bynderClient.DownloadAsset(asset.Id);
-            if (file == null)
+            if (asset.Type == AssetType.Video)
             {
-                _inRiverContext.Log(LogLevel.Error, $"Could not download asset with Id {asset.Id}");
-
-                bynderDownloadStateField.Data = BynderStates.Error;
-                _inRiverContext.ExtensionManager.DataService.UpdateFieldsForEntity(new List<Field> { bynderDownloadStateField });
-
-                return;
-            }
-
-            if (existingFileId > 0)
-            {
-                // check if existing ResourceFile is same size as the one in Bynder
-                var resourceFileMetaData = _inRiverContext.ExtensionManager.UtilityService.GetFileMetaData(existingFileId);
-                if (file.LongLength == resourceFileMetaData.FileSize)
+                if (asset.Thumbnails.TryGetValue("webimage", out var previewUrl))
                 {
-                    _inRiverContext.Log(LogLevel.Debug, "Asset is the same as before. Setting the state to Done.");
-                    bynderDownloadStateField.Data = BynderStates.Done;
-                    _inRiverContext.ExtensionManager.DataService.UpdateFieldsForEntity(new List<Field> { bynderDownloadStateField });
+                    var fileId = _inRiverContext.ExtensionManager.UtilityService.AddFileFromUrl(asset.GetOriginalFileName(), previewUrl);
+                    resourceEntity.GetField(FieldTypeIds.ResourceFileId).Data = fileId;
+                }
+            }
+            else
+            {
+                var file = _bynderClient.DownloadAsset(asset.Id);
+                if (file == null)
+                {
+                    _inRiverContext.Log(LogLevel.Error, $"Could not download asset with Id {asset.Id}");
+
+                    bynderDownloadStateField.Data = BynderStates.Error;
+                    _inRiverContext.ExtensionManager.DataService.UpdateFieldsForEntity(new List<Field> {bynderDownloadStateField});
+
                     return;
                 }
-                // delete older asset file
-                _inRiverContext.Log(LogLevel.Verbose, $"existing fileId found {existingFileId}");
-                _inRiverContext.ExtensionManager.UtilityService.DeleteFile(existingFileId);
+
+                if (existingFileId > 0)
+                {
+                    // check if existing ResourceFile is same size as the one in Bynder
+                    var resourceFileMetaData =
+                        _inRiverContext.ExtensionManager.UtilityService.GetFileMetaData(existingFileId);
+                    if (file.LongLength == resourceFileMetaData.FileSize)
+                    {
+                        _inRiverContext.Log(LogLevel.Debug, "Asset is the same as before. Setting the state to Done.");
+                        bynderDownloadStateField.Data = BynderStates.Done;
+                        _inRiverContext.ExtensionManager.DataService.UpdateFieldsForEntity(new List<Field> {bynderDownloadStateField});
+                        return;
+                    }
+
+                    // delete older asset file
+                    _inRiverContext.Log(LogLevel.Verbose, $"existing fileId found {existingFileId}");
+                    _inRiverContext.ExtensionManager.UtilityService.DeleteFile(existingFileId);
+                }
+
+                // add new asset
+                string resourceFileName = (string) resourceEntity.GetField(FieldTypeIds.ResourceFilename)?.Data;
+                int newFileId = _inRiverContext.ExtensionManager.UtilityService.AddFile(resourceFileName, file);
+
+                // set fieltypes for resource entity
+                resourceEntity.GetField(FieldTypeIds.ResourceFileId).Data = newFileId;
             }
 
-            // add new asset
-            string resourceFileName = (string)resourceEntity.GetField(FieldTypeIds.ResourceFilename)?.Data;
-            int newFileId = _inRiverContext.ExtensionManager.UtilityService.AddFile(resourceFileName, file);
-
-            // set fieltypes for resource entity
-            resourceEntity.GetField(FieldTypeIds.ResourceFileId).Data = newFileId;
             resourceEntity.GetField(FieldTypeIds.ResourceMimeType).Data = asset.GetOriginalMimeType();
+            
             bynderDownloadStateField.Data = BynderStates.Done;
 
             _inRiverContext.ExtensionManager.DataService.UpdateEntity(resourceEntity);
